% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{ambiguous_assignment}
\alias{ambiguous_assignment}
\title{Identify ambiguous cluster assignment}
\usage{
ambiguous_assignment(x)
}
\arguments{
\item{x}{Numeric matrix of distances}
}
\value{
Logical vector corresponding to results on each row
}
\description{
Identify positions where the difference between the minimum distance to a cluster and the distance any other is
less than 0.03, which indicates ambiguous assignment.
}
\details{
Internal function called by \code{\link{annotate_df}}.
}
\keyword{internal}
