% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_df}
\alias{annotate_df}
\title{Assign PCA, UMAP and clusters to a data frame of positional ER scores}
\usage{
annotate_df(df)
}
\arguments{
\item{df}{Data frame}
}
\value{
A list containing two data frames, the first covering PCA and UMAP assignment and the second
cluster assignment
}
\description{
Internal function called by \code{\link{annotate}} to perform the mechanics of annotations
}
\keyword{internal}
