% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_dms}
\alias{annotate_dms}
\title{Annotate a Deep Mutational Scan object}
\usage{
annotate_dms(x)
}
\arguments{
\item{x}{\code{\link{deep_mutational_scan}}}
}
\description{
Internal function called by \code{\link{annotate}} when passed a \code{\link{deep_mutational_scan}}
}
