% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_dms_deep_scan}
\alias{annotate_dms_deep_scan}
\title{Annotate a Deep Mutational Scan object}
\usage{
annotate_dms_deep_scan(x)
}
\arguments{
\item{x}{\code{\link{deep_mutational_scan}}}
}
\description{
Internal function called by \code{\link{annotate_dms}} when passed a \code{\link{deep_mutational_scan}}
}
