% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_dms_tibble}
\alias{annotate_dms_tibble}
\title{Annotate a Deep Mutational Scan object}
\usage{
annotate_dms_tibble(x)
}
\arguments{
\item{x}{\code{\link[tibble]{tibble}}}
}
\description{
Internal function called by \code{\link{annotate_dms}} when passed a \code{\link[tibble]{tibble}}
}
