% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{any_less_than}
\alias{any_less_than}
\title{Check if any non-zero values are below a threshold}
\usage{
any_less_than(x, threshold)
}
\arguments{
\item{x}{Numeric vector}

\item{threshold}{Threshold below which TRUE is returned}
}
\description{
Check if any non-zero values are below a threshold
}
\keyword{internal}
