% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{as_tibble}
\alias{as_tibble}
\alias{as_tibble.deep_mutational_scan}
\alias{as.data.frame.deep_mutational_scan}
\title{Convert deep_mutational_scans to data frames}
\usage{
\method{as_tibble}{deep_mutational_scan}(x, ..., full = FALSE)

\method{as.data.frame}{deep_mutational_scan}(x, ..., full = FALSE)
}
\arguments{
\item{x}{\link{deep_mutational_scan} object}

\item{...}{Ignored}

\item{full}{Include columns containing metadata (study, gene)}
}
\description{
Convert deep_mutational_scans to data frames
}
\section{Methods (by class)}{
\itemize{
\item \code{deep_mutational_scan}: S3 as.data.frame method
}}

