% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{autoplot.deep_mutational_scan}
\alias{autoplot.deep_mutational_scan}
\title{Autoplot method for Deep Mutational Scans}
\usage{
\method{autoplot}{deep_mutational_scan}(x, ...)
}
\description{
Autoplot method for Deep Mutational Scans
}
