% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{bind_scans}
\alias{bind_scans}
\title{Combine deep mutational scan data}
\usage{
bind_scans(..., impute_missing = FALSE, annotate_missing = FALSE)
}
\arguments{
\item{...}{\code{\link{deep_mutational_scan}} objects to combine or lists of such objects.}

\item{impute_missing}{Impute missing ER scores using \code{\link{impute}} where scans have not already been imputed.}

\item{annotate_missing}{Annotate unannotated scans using \code{\link{annotate}}.}
}
\value{
A multi study \code{\link{deep_mutational_scan}}.
}
\description{
Combine multiple \code{\link{deep_mutational_scan}} objects into a single \code{\link[tibble]{tibble}}, including
columns giving the study and gene of each position.
}
