% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{calculate_margin}
\alias{calculate_margin}
\title{Calculate the margin by which each element is largest than the smallest}
\usage{
calculate_margin(x)
}
\arguments{
\item{x}{Numeric vector}
}
\description{
Calculate the margin by which each element is largest than the smallest
}
