% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_data}
\alias{check_data}
\title{Check a deep_mutational_scan for common anomalies}
\usage{
check_data(x, warn = TRUE)
}
\arguments{
\item{x}{\code{\link{deep_mutational_scan}}.}

\item{warn}{Produce warnings when anomalies are detected.}
}
\value{
Character vector of anomalies detected
}
\description{
Identify common potential anomalies in deep mutational scanning data, for example an unusual distribution of ER
scores or unexpected cluster assignment. Some checks require the data to be annotated and will be skipped, with a
warning, if it is not.
}
\details{
The following checks are performed:
\itemize{
  \item Median ER is positive, suggesting it has possibly been inverted.
}

If \code{x} is annotated the following additional checks are performed:
\itemize{
  \item There's an unusually low number of common subtypes (1 and 2), suggesting unusual positions or an error like
  an incorrect sequence.
}
}
