% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{check_permissive}
\alias{check_permissive}
\title{Check if ER scores qualifies as permissive}
\usage{
check_permissive(x)
}
\arguments{
\item{x}{Vector of ER scores#}
}
\description{
Check if ER scores qualifies as permissive
}
\keyword{internal}
