% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clamp}
\alias{clamp}
\title{Clamp values to limits}
\usage{
clamp(x, lower = -Inf, upper = Inf)
}
\arguments{
\item{x}{Numeric vector to clamp}

\item{upper, lower}{Limits to clamp to}
}
\description{
Clamp values to limits
}
