% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{cluster_notes}
\alias{cluster_notes}
\title{Generate cluster notes message}
\usage{
cluster_notes(permissive, ambiguous, high_distance)
}
\arguments{
\item{permissive}{Logical showing which positions are permissive}

\item{ambiguous}{Logical showing which positions are ambiguous}

\item{high_distance}{Logical showing which positions are outliers due to distance from all centroids.}
}
\value{
Character vector
}
\description{
Create a standard note for cluster assignments, explaining why O, P or A cluster status is assigned.
}
\details{
Internal function called by \code{\link{annotate_df}}.
}
\keyword{internal}
