% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare_dms}
\alias{compare_dms}
\title{Compare a new DMS dataset to the combined landscape}
\usage{
compare_dms(x)
}
\arguments{
\item{x}{\link{deep_mutational_scan} to analyse}
}
\description{
Compare a new DMS dataset to the combined landscape
}
