% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{cosine_distance}
\alias{cosine_distance}
\title{Calculate Cosine distance between rows of two matrices}
\usage{
cosine_distance(x, y)
}
\arguments{
\item{x, y}{Numeric matrices}
}
\description{
Calculate Cosine distance between rows of two matrices
}
