% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recluster.R
\name{cosine_distance_matrix}
\alias{cosine_distance_matrix}
\title{Cosine distance matrix}
\usage{
cosine_distance_matrix(m)
}
\arguments{
\item{m}{Numeric matrix}
}
\description{
Calculate the cosine distance between all rows of a matrix.
}
