% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deep_mutational_scans}
\alias{deep_mutational_scans}
\title{Standardised deep mutational scanning studies results}
\format{
A data frame with 6357 rows
\describe{
  \item{study}{Source study}
}
}
\source{
\href{https://www.biorxiv.org/content/10.1101/2020.05.26.116756v1}{Dunham and Beltrao (2020)}
}
\usage{
deep_mutational_scans
}
\description{
A dataset containing combined results from 28 deep mutational scanning
studies. The fitness scores are transformed into the same format and
normalised.
}
\keyword{datasets}
