% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deep_scan}
\alias{deep_scan}
\title{Example deep_mutational_scan}
\format{
A deep_mutational_scan with 306 positions
\describe{
  \item{data}{Positional data}
  \item{study}{Name/ID of the source study}
  \item{gene}{Gene scanned}
  \item{annotated}{Logical indicating if the scan has been annotated with deep landscape data}
  \item{impute_mask}{Numeric matrix showing which ER scores were imputed}
  \item{cluster}{Tibble detailing amino acid subtype assignment, or NA}
}
}
\usage{
deep_scan
}
\description{
A \code{\link{deep_mutational_scan}} containing data from XXX. This dataset is primarily used to illustrate use of
the package functions.
}
\keyword{data}
