% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{deep_scans}
\alias{deep_scans}
\title{Example deep_mutational_scan objects}
\format{
A list of \code{\link{deep_mutational_scan}} objects
\describe{
  \item{hsp90}{HSP90 scan by Hietpas et al. (2011, doi: 10.1073/pnas.1016024108, MaveDB: urn:mavedb:00000011-a-1)}
  \item{gpa}{GpA scan by Elazar et al. (2016, doi: 10.7554/eLife.12125, MaveDB:  urn:mavedb:00000051-c-1)}
  \item{p53}{p53 scan by Kotler et al. (2018, doi: 10.1016/j.molcel.2018.06.012 , MaveDB: urn:mavedb:00000059-c-1)}
}
}
\usage{
deep_scans
}
\description{
A list of \code{\link{deep_mutational_scan}} objects. This dataset is primarily used to illustrate use of
the package functions. The studies were sourced from \url{https://www.mavedb.org}{MaveDB}.
}
\keyword{data}
