% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{dms_extract}
\alias{dms_extract}
\alias{[.deep_mutational_scan}
\alias{[<-.deep_mutational_scan}
\alias{[[.deep_mutational_scan}
\alias{[[<-.deep_mutational_scan}
\title{Extracting and replacing deep mutational scanning data}
\usage{
\method{[}{deep_mutational_scan}(x, i, j, drop = FALSE, ...)

\method{[}{deep_mutational_scan}(x, i, j, ...) <- value

\method{[[}{deep_mutational_scan}(x, i, j, exact = FALSE, ...)

\method{[[}{deep_mutational_scan}(x, i, j, ...) <- value
}
\arguments{
\item{x}{\link{deep_mutational_scan} object}

\item{i, j, ...}{Indeces to access}

\item{drop}{Coerce result to lowest possible dimension}

\item{value}{Value to set}

\item{exact}{Ignored for tibbles (which is the underlying data structure)}
}
\description{
Extracting and replacing deep mutational scanning data
}
\section{Functions}{
\itemize{
\item \code{[.deep_mutational_scan}: Extract

\item \code{[<-.deep_mutational_scan}: Assign

\item \code{[[.deep_mutational_scan}: Double Extract

\item \code{[[<-.deep_mutational_scan}: Double Assign
}}

