% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{dms_plot}
\alias{dms_plot}
\alias{autoplot.deep_mutational_scan}
\alias{plot.deep_mutational_scan}
\title{Summary plot for Deep Mutational Scans}
\usage{
\method{autoplot}{deep_mutational_scan}(object, ...)

\method{plot}{deep_mutational_scan}(x, ...)
}
\arguments{
\item{object, x}{\link{deep_mutational_scan} to produce a
summary plot for.}

\item{...}{Additional parameters}
}
\description{
Summary plot for Deep Mutational Scans
}
\section{Functions}{
\itemize{
\item \code{plot.deep_mutational_scan}: S3 plot method
}}

