% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{dms_s3}
\alias{dms_s3}
\alias{format.deep_mutational_scan}
\alias{print.deep_mutational_scan}
\alias{str.deep_mutational_scan}
\alias{dim.deep_mutational_scan}
\title{General S3 Methods for Deep Mutational Scans}
\usage{
\method{format}{deep_mutational_scan}(x, ...)

\method{print}{deep_mutational_scan}(x, ...)

\method{str}{deep_mutational_scan}(object, ..., indent.str = " ", nest.lev = 0)

\method{dim}{deep_mutational_scan}(x)
}
\arguments{
\item{x, object}{\link{deep_mutational_scan} object}

\item{...}{Additional arguments}

\item{indent.str}{Indent string for \link[utils]{str}}

\item{nest.lev}{Nest level for \link[utils]{str}}
}
\description{
General S3 Methods for Deep Mutational Scans
}
\section{Functions}{
\itemize{
\item \code{format.deep_mutational_scan}: S3 format method

\item \code{print.deep_mutational_scan}: S3 print method

\item \code{str.deep_mutational_scan}: S3 str method

\item \code{dim.deep_mutational_scan}: S3 dim method
}}

