% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{dms_summary}
\alias{dms_summary}
\alias{summary.deep_mutational_scan}
\alias{format.deep_mutational_scan_summary}
\alias{print.deep_mutational_scan_summary}
\title{Summarise Deep Mutational Scans}
\usage{
\method{summary}{deep_mutational_scan}(object, ...)

\method{format}{deep_mutational_scan_summary}(x, ...)

\method{print}{deep_mutational_scan_summary}(x, ...)
}
\arguments{
\item{object, x}{A \link{deep_mutational_scan} object.}

\item{...}{Additional arguments}
}
\value{
deep_mutational_scan_summary object containing:
  \describe{
    \item{study}{The source study}
    \item{gene}{The gene scanned}
    \item{annotated}{Logical showing whether the dataset has been annotated with deep mutational landscape data}
    \item{positions}{Number of positions with data}
    \item{er}{Named vector containing summary statistics to ER scores}
    \item{imputed}{Named vector summarising the number of imputed scores}
    \item{clusters}{Named vector summarising cluster annotations}
  }
}
\description{
Summarise Deep Mutational Scans
}
\section{Functions}{
\itemize{
\item \code{format.deep_mutational_scan_summary}: S3 format method

\item \code{print.deep_mutational_scan_summary}: S3 print method
}}

