% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{get_feature_scale}
\alias{get_feature_scale}
\title{Select colour scales for the features in the combined dataset}
\usage{
get_feature_scale(feature, type = c("fill", "colour"))
}
\arguments{
\item{feature}{Feature to fetch the scale for.}

\item{type}{Return a fill or a colour scale}
}
\description{
Select from preconfigured scales for the various features in the combined dataset
}
\keyword{internal}
