% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_dms}
\alias{impute_dms}
\title{Impute missing deep mutational scan data}
\usage{
impute_dms(x, na_value = "impute")
}
\arguments{
\item{x}{link{deep_mutational_scan} to impute data from}

\item{na_value}{Value to set missense NA values to. Setting this to "impute" sets scores to the median score for that
substitution (e.g A -> C) in the combined dataset}
}
\description{
Impute missing deep mutational scan data
}
