% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{is.deep_mutational_scan}
\alias{is.deep_mutational_scan}
\title{Determine if an object is a deep_mutational_scan}
\usage{
is.deep_mutational_scan(x)
}
\arguments{
\item{x}{Object to check}
}
\description{
Determine if an object is a deep_mutational_scan
}
