% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{landscape_outliers}
\alias{landscape_outliers}
\title{Identify positions outside the mutational landscape}
\usage{
landscape_outliers(x, cutoff = 0.25)
}
\arguments{
\item{x}{Annotated \code{\link{deep_mutational_scan}}.}

\item{cutoff}{Minimum distance from any studied position to be labelled an outlier.}
}
\value{
integer vector containing the rows of outlying positions
}
\description{
Find positions that lie outside the studied deep mutational landscape.
}
\examples{
x <- annotate(deep_scans$p53)
landscape_outliers(x)
}
