% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{landscape_properties}
\alias{landscape_properties}
\title{Predict positions properties from their position in the deep mutational landscape}
\usage{
landscape_properties(x, bins = 20)
}
\arguments{
\item{x}{\link{deep_mutational_scan}.}

\item{bins}{Number of bins to divide landscape into. The default value was chosen to give the best results on a
benchmark dataset.}
}
\value{
A \code{\link[tibble]{tibble}}, with each row detailing a row of the input data and columns matching
those in the \code{\link{subtypes}} dataset, describing the predicted properties of the position.
}
\description{
Predict positions properties based on the average properties of \code{\link{deep_landscape}} positions in the same
hexagonal bin as them.
}
\details{
This function is not externally exposed or used for other annotation because subtype based annotation is found to be
at least as accurate in almost all cases, and often better. It is included for illustration and potential further
development work.
}
\keyword{internal}
