% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{new_deep_mutational_scan}
\alias{new_deep_mutational_scan}
\title{Internal deep_mutational_scan constructor}
\usage{
new_deep_mutational_scan(df, study, gene)
}
\arguments{
\item{df}{A data frame in the required format (position, wt, A-Y, additional data...).}

\item{study}{The source study.}

\item{gene}{The studied gene.}
}
\value{
A deep_mutational_scan S3 object (see \code{\link{deep_mutational_scan}})
}
\description{
Create a new raw deep_mutational_scan object. This is expected to have no annotations and no imputation but have
had scores transformed and normalised onto the standard scale.
}
