% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise.R
\name{normalise_er}
\alias{normalise_er}
\title{Normalise deep mutational scanning fitness scores}
\usage{
normalise_er(x, q = 0.1)
}
\arguments{
\item{x}{Vector of fitness scores}

\item{q}{Quantile to normalise against. Values other than 0.1 differ from
those used in the combined dataset}
}
\value{
Numeric vector of normalised ER scores
}
\description{
Normalise log2(ER) scores by dividing my the median of the lowest 10% of score, under the assumption these will
represent complete protein knockout and will therefore be somewhat comparable between studies.
}
