% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{outlier_positions}
\alias{outlier_positions}
\title{Identify outlier positions}
\usage{
outlier_positions(x)
}
\arguments{
\item{x}{Numeric matrix of distances}
}
\value{
Logical vector corresponding to results on each row
}
\description{
Identify positions where the distance to all clusters is > 0.45, indicating a likely outlier.
}
\details{
Internal function called by \code{\link{annotate_df}}.
}
\keyword{internal}
