% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_deep_scan}
\alias{parse_deep_scan}
\title{Parse deep mutational scanning data}
\usage{
parse_deep_scan(x, scheme, position_offset = 0, score_col = NULL)
}
\arguments{
\item{x}{A tibble or an object that can be converted to one}

\item{scheme}{Original data scheme (see description). Accepts any unambiguous
substring.}

\item{position_offset}{Offset all positions by this amount}

\item{score_col}{Name of column containing fitness scores (used for when
parsing MaveDB data)}
}
\description{
Parse deep mutational scanning data
}
