% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_dms}
\alias{parse_dms}
\title{Parse deep mutational scanning data}
\usage{
parse_dms(x, path)
}
\description{
Parse deep mutational scanning data from files or R objects, converting it
into a format that deepscanscape can opperate on.
}
