% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{permissive_positions}
\alias{permissive_positions}
\title{Identify permissive positions}
\usage{
permissive_positions(x)
}
\arguments{
\item{x}{Numeric matrix of ER scores}
}
\value{
Logical vector corresponding to results on each row
}
\description{
Identify rows of a matrix of ER scores where all scores are less than 0.4 in absolute value
}
\keyword{internal}
