% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{plot.deep_mutational_scan}
\alias{plot.deep_mutational_scan}
\title{Plot method for Deep Mutational Scans}
\usage{
\method{plot}{deep_mutational_scan}(x, ...)
}
\description{
Plot method for Deep Mutational Scans
}
