% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_cluster_frequencies}
\alias{plot_cluster_frequencies}
\title{Plot amino acid subtype frequencies}
\usage{
plot_cluster_frequencies(x)
}
\arguments{
\item{x}{\code{\link{deep_mutational_scan}}.}
}
\value{
A \code{\link[ggplot2]{ggplot2}} plot.
}
\description{
Plot the frequency each cluster occurs in an annotated deep mutational scan dataset, in order to identify data that
strays from the expected frequencies. For example large proportions of outliers could suggest abnormal data or many
permissive positions a weakly conserved protein.
}
\examples{
dms <- annotate(deepscanscape::deep_scans$p53)
plot_cluster_frequencies(dms)

# Plot multiple studies
comb_dms <- bind_scans(dms, annotate_missing = TRUE)
plot_cluster_frequencies(comb_dms)

}
