% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_dms_heatmap}
\alias{plot_dms_heatmap}
\title{Plot a deep mutational scan dataset as a heatmap}
\usage{
plot_dms_heatmap(x)
}
\arguments{
\item{x}{\link{deep_mutational_scan} to analyse}
}
\description{
Plot a deep mutational scan dataset as a heatmap
}
