% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recluster.R
\name{plot_dms_recluster}
\alias{plot_dms_recluster}
\title{Summarise a deep mutational scan recluster}
\usage{
plot_dms_recluster(df, aa = NULL, square_tiles = TRUE)
}
\arguments{
\item{df}{reclustered data, as output by \link{recluster_dms}}

\item{aa}{Character vector of amino acids to summarise clusters from.}

\item{square_tiles}{Force heatmap tiles to be square. It can be useful to disable this when summarising a large
number of clusters.}
}
\description{
Summarise a deep mutational scan recluster
}
