% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_er_distribution}
\alias{plot_er_distribution}
\title{Plot ER distributions}
\usage{
plot_er_distribution(x)
}
\arguments{
\item{x}{\code{\link{deep_mutational_scan}}.}
}
\value{
A \code{\link[ggplot2]{ggplot2}} plot.
}
\description{
Compare the ER score distribution from a collection of new studies to those in the \link{deep_landscape} dataset.
}
\examples{
# Plot a single studies ER
dms <- deepscanscape::deep_scans$p53
plot_er_distribution(dms)

# Plot multiple studies
comb_dms <- bind_scans(dms, annotate_missing = TRUE)
plot_er_distribution(comb_dms)
}
