% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_er_heatmap}
\alias{plot_er_heatmap}
\title{Plot a heat map of ER scores from a deep mutational scan}
\usage{
plot_er_heatmap(x)
}
\arguments{
\item{x}{\code{\link{deep_mutational_scan}}.}
}
\value{
A \code{\link[ggplot2]{ggplot2}} plot.
}
\description{
Plot a heat map showing the ER fitness score for each mutation at each position and the positions mean ER score.
The wild type amino acid at each position is marked by outlining the corresponding heatmap tile.
This is also the default plot method for \code{\link{deep_mutational_scan}} objects.
}
\examples{
dms <- deepscanscape::deep_scans$p53
plot_er_heatmap(dms)

# Or equivalently
plot(dms)

}
