% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_landscape_plain}
\alias{plot_landscape_plain}
\title{Plot new data onto the plain deep mutational landscape}
\usage{
plot_landscape_plain(x)
}
\arguments{
\item{x}{\code{\link{deep_mutational_scan}}.}
}
\value{
A \code{\link[ggplot2]{ggplot2}} plot.
}
\description{
Internal helper called by plot_landscape when no feature is passed
}
\keyword{internal}
