% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{rbind.deep_mutational_scan}
\alias{rbind.deep_mutational_scan}
\title{Combine deep mutational scan data}
\usage{
\method{rbind}{deep_mutational_scan}(...)
}
\arguments{
\item{...}{\link{deep_mutational_scan} objects to combine}
}
\description{
Combine deep mutational scan data
}
