% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise.R
\name{standardise_dms}
\alias{standardise_dms}
\title{Again, could work on df or vector, probably df
Standardise DMS data}
\usage{
standardise_dms(x)
}
\description{
Again, could work on df or vector, probably df
Standardise DMS data
}
