% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{summary.deep_mutational_scan}
\alias{summary.deep_mutational_scan}
\title{Summary method for Deep Mutational Scans}
\usage{
\method{summary}{deep_mutational_scan}(x, ...)
}
\description{
Summary method for Deep Mutational Scans
}
