% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{theme_deepscanscape}
\alias{theme_deepscanscape}
\title{Default deepscanscape GGPlot2 theme}
\usage{
theme_deepscanscape()
}
\description{
A GGPlot2 theme based on \code{\link[ggpubr]{theme_pubclean}}, with a generally minimalist style. This theme
changes the legend to be on the right, has a central title and subtitle, and removes facet strip and legend
key backgrounds.
}
