% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{validate_combined_dms}
\alias{validate_combined_dms}
\title{Check a data frame is a combined mutational scan dataset}
\usage{
validate_combined_dms(x, annotated = FALSE)
}
\arguments{
\item{x}{Data frame to check.}

\item{annotated}{logical indicating the data be annotated with PCA, UMAP and Cluster information.}
}
\value{
x converted to a \code{\link[tibble]{tibble}}
}
\description{
Validate a data frame to determine if it sufficiently resembles a combined deep mutational scan dataset, as produced
by \code{\link{rbind.deep_mutational_scan}}. A error is raised if the data is incorrect, otherwise it is returned
as a \code{\link[tibble]{tibble}}.
}
\details{
The following tests are performed:
\itemize{
  \item x is a \code{\link{data.frame}}
  \item There are no NA ER scores
  \item The required columns are present
  \item The expected additional columns are present for annotated data
}
}
