% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{validate_deep_mutational_scan}
\alias{validate_deep_mutational_scan}
\title{Validate deep_mutational_scan objects}
\usage{
validate_deep_mutational_scan(x)
}
\arguments{
\item{x}{\link{deep_mutational_scan} object}
}
\description{
Validate deep_mutational_scan objects
}
