% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_df}
\alias{annotate_df}
\title{Annotate a Deep Mutational Scan object}
\usage{
annotate_df(x)
}
\arguments{
\item{x}{\code{\link[tibble]{tibble}}}
}
\description{
Internal function called by \code{\link{annotate}} when passed a \code{\link[tibble]{tibble}}
}
\keyword{internal}
