% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_dms}
\alias{annotate_dms}
\title{Annotate positions based on the deep mutational landscape}
\usage{
annotate_dms(x)
}
\arguments{
\item{x}{\code{\link{deep_mutational_scan}} to annotate}
}
\description{
Annotate positions based on the deep mutational landscape
}
