% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{as_tibble}
\alias{as_tibble}
\alias{as_tibble.deep_mutational_scan}
\alias{as.data.frame.deep_mutational_scan}
\alias{as.list.deep_mutational_scan}
\title{Convert deep_mutational_scan objects}
\usage{
\method{as_tibble}{deep_mutational_scan}(x, ..., full = FALSE)

\method{as.data.frame}{deep_mutational_scan}(x, ..., full = FALSE)

\method{as.list}{deep_mutational_scan}(x, ...)
}
\arguments{
\item{x}{\code{\link{deep_mutational_scan}} object.}

\item{...}{Ignored.}

\item{full}{Include columns containing metadata (study, gene).}
}
\value{
A \code{\link[tibble]{tibble}}, \code{\link[base]{data.frame}} or \code{\link[base]{list}}
}
\description{
Conveniently convert \code{\link{deep_mutational_scan}} objects to a \code{\link[tibble]{tibble}} or
\code{\link[base]{data.frame}} or \code{\link[base]{list}}, with the option to add gene and study columns.
}
\section{Methods (by class)}{
\itemize{
\item \code{deep_mutational_scan}: S3 as.data.frame method

\item \code{deep_mutational_scan}: S3 as.list method
}}

