% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{calculate_cluster_distances}
\alias{calculate_cluster_distances}
\title{Calculate distance to each cluster of the amino acid}
\usage{
calculate_cluster_distances(wt, pca)
}
\arguments{
\item{wt}{Wild type amino acid}

\item{pca}{PCA values for PC2 to PC20}
}
\description{
Calculate distance to each cluster of the amino acid
}
\keyword{internal}
