% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check_data}
\alias{check_data}
\title{Check a deep_mutational_scan for common anomalies}
\usage{
check_data(x, warn = TRUE)
}
\arguments{
\item{x}{\code{\link{deep_mutational_scan}}.}

\item{warn}{Produce warnings when anomalies are detected.}
}
\value{
Data frame of anomalies detected
}
\description{
Identify common potential anomalies in deep mutational scanning data, for example an unusual distribution of ER
scores or unexpected cluster assignment. Some checks require the data to be annotated and will be skipped, with a
warning, if it is not.
}
\details{
The following checks are performed:
\itemize{
  \item Median ER is positive, suggesting it has possibly been inverted.
  \item Mean WT |ER| > 0.1, suggesting an incorrect sequence.
}

If \code{x} is annotated the following additional checks are performed:
\itemize{
  \item Positions fall outside the mutational landscape, suggesting they may have unusual function or the experiment
  selects unusual properties.
  \item There are an unusually high number of permissive positions, based on a binomial test. This can suggest a
  weakly selective region or experiment.
}
}
\examples{
check_data(deep_scans$p53, warn = TRUE)
}
