% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cosine_distance_matrix}
\alias{cosine_distance_matrix}
\title{Cosine distance matrix}
\usage{
cosine_distance_matrix(x, y = NULL)
}
\arguments{
\item{x, y}{Numeric matrices. The distance between rows of x is calculated if y = NULL.}
}
\description{
Calculate the cosine distance between rows of a matrix or matrices.
}
\details{
Floating point imprecision in the implementation of \code{\link[base]{tcrossprod}} leads to values slightly outside
the domain of acos, so cosine similarities are rounded to 8 digits before being converted to distances. This is well
beyond the precision required for most applications and does not pose a problem, especially here where the
experimental uncertainty of ER scores are greater than this.
}
\keyword{internal}
