% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{deep_mutational_scan}
\alias{deep_mutational_scan}
\title{Deep mutational scan data}
\usage{
deep_mutational_scan(
  df,
  scheme = NULL,
  trans = NULL,
  na_value = "impute",
  annotate = TRUE,
  study = NA,
  gene = NA
)
}
\arguments{
\item{df}{A data frame to parse.}

\item{scheme}{Original data scheme (see \code{\link{parse_deep_scan}}).}

\item{trans}{Function to transform scores onto the standard scale. Accepts a string corresponding to known transforms
or a custom function (\code{\link{transform_er}}).}

\item{na_value}{Value to set missense NA scores to (see \code{\link{impute}}).}

\item{annotate}{Annotate the dataset with mutational landscape data (PCA, UMAP and amino acid subtypes).}

\item{study}{Source of the deep mutational scan.}

\item{gene}{Gene scanned.}
}
\value{
A deep_mutational_scan S3 object, which broadly behaves as a list containing the following fields:
  \itemize{
    \item data: \code{\link[tibble]{tibble}} containing positions ER scores and other positional data
    \item study: source study
    \item gene: gene scanned
    \item annotated: logical indicating if the data has been annotated with PCs, UMAP coordinates and clusters
    \item impute_mask: numeric matrix indicating imputed ER scores, corresponding to the
    rows/ER columns of \code{data} (see \code{\link{impute}} for details)
    \item cluster: \code{\link[tibble]{tibble}} containing further details on how clusters were assigned during
    annotation, corresponding to the rows of \code{data}
  }
  The class is used like a list apart from [ accesses the main data frame (\link[=dms_extract]{details}).
  The class is also associated with other common generics (see \link[=dms_s3]{basic functions},
  \link[=dms_summary]{summary}, \link[=as_tibble]{data frames}, \link[=dms_plot]{plotting} and
  \link[=rbind.deep_mutational_scan]{combining data}).
}
\description{
Store deep mutational scanning data in a standardised format, alongside
metadata describing the state of the data.
}
\examples{
path <- system.file("extdata", "urn_mavedb_00000011_a_1_scores.csv",
                    package = "deepscanscape")
csv <- read.csv(path, skip = 4)
dms <- deep_mutational_scan(csv, scheme = 'mave', trans = NULL, na_value = 'impute',
                            annotate = FALSE, gene = 'Hsp90', study = 'Hietpas et al. (2011)')
}
