% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{deep_mutational_scan}
\alias{deep_mutational_scan}
\title{Deep mutational scan data}
\usage{
deep_mutational_scan(
  df,
  scheme = NULL,
  trans = NULL,
  na_value = "impute",
  annotate = TRUE,
  study = NULL,
  gene = NULL
)
}
\arguments{
\item{df}{A data frame to parse}

\item{scheme}{Original data scheme (see \code{\link{parse_dms_data}})}

\item{trans}{Function to transform scores onto the standard scale. Accepts a string corresponding to known transforms
or a custom function (\code{\link{transform_dms}}).}

\item{na_value}{Value to set missense NA scores to (see \code{\link{impute_dms}})}

\item{annotate}{Annotate the dataset with mutational landscape data (PCA, UMAP and amino acid subtypes)}

\item{study}{Source of the deep mutational scan}

\item{gene}{Gene scanned}
}
\value{
A deep mutational scan S3 object
}
\description{
Store deep mutational scanning data in a standardised format, alongside
metadata describing the state of the data. The primary data is contained in
a tibble
}
