% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{deep_mutational_scan}
\alias{deep_mutational_scan}
\title{Deep mutational scan data}
\usage{
deep_mutational_scan(x)
}
\arguments{
\item{x}{A data frame to parse}

\item{format}{Original data format (see description)}
}
\value{
A deep mutational scan S3 object
}
\description{
Store deep mutational scanning data in a standardised format, alongside
metadata describing the state of the data. The primary data is contained in
a data frame
}
