% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deepscanscape.R
\docType{package}
\name{deepscanscape}
\alias{deepscanscape}
\title{deepscanscape: Check and Annotate Deep Mutational Scanning Data}
\description{
The deepscanscape package provides functions to standardise deep
mutational scanning data and compare it to a deep mutational landscape
calculated from a dataset of 28 previous studies. This allows the whole
dataset to be checked for unusual properties and individual positions
to be annotated with positional subtypes, indicating the positions in other
proteins they are most similar to.
}
\section{Functions}{
NA
}

