% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotate.R
\name{describe_clusters}
\alias{describe_clusters}
\title{Describe amino acid positional subtypes}
\usage{
describe_clusters(x)
}
\arguments{
\item{x}{\link{deep_mutational_scan}}
}
\description{
Describe amino acid positional subtypes
}
