% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{dms_extract}
\alias{dms_extract}
\alias{[.deep_mutational_scan}
\alias{[<-.deep_mutational_scan}
\title{Extracting and replacing deep mutational scanning data}
\usage{
\method{[}{deep_mutational_scan}(x, i, j, drop = FALSE, ...)

\method{[}{deep_mutational_scan}(x, i, j, ...) <- value
}
\arguments{
\item{x}{\code{\link{deep_mutational_scan}} object.}

\item{i, j, ...}{Indices to access.}

\item{drop}{Coerce result to lowest possible dimension.}

\item{value}{Value to set.}
}
\description{
Extracting and replacing data from \code{\link{deep_mutational_scan}} objects uses a mixture of list and data frame
syntax. $ and [[ extract values from the main data fields (e.g. the gene or where the data is annotated) and [
provides a shortcut to access the main data table, which can otherwise be accessed via \code{x$data}.
}
\section{Functions}{
\itemize{
\item \code{[.deep_mutational_scan}: Extract

\item \code{[<-.deep_mutational_scan}: Assign
}}

\examples{
# Setup object
path <- system.file("extdata", "urn_mavedb_00000036-a-1_scores.csv",
                    package = "deepscanscape")
csv <- read.csv(path, skip = 4)
dms <- deep_mutational_scan(csv, scheme = "mave", trans = "vamp",
                            gene = "LDLRAP1", study = "urn:mavedb:00000036")

# Extract meta data:
dms$study
dms$impute_mask

dms[["gene"]]

# Replace meta data
dms$gene <- "new_gene"
dms$impute_mask <- NA

# Quickly access data columns:
dms["A"]
dms[c(1,2,3), "C"]
dms[c("position", "wt", "A", "D")]

# Quickly modify data columns
dms["position"] <- dms["position"] + 1

}
