% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{dms_plot}
\alias{dms_plot}
\alias{autoplot.deep_mutational_scan}
\alias{plot.deep_mutational_scan}
\title{Summary plot for Deep Mutational Scans}
\usage{
\method{autoplot}{deep_mutational_scan}(object, ...)

\method{plot}{deep_mutational_scan}(x, ...)
}
\arguments{
\item{object, x}{\code{\link{deep_mutational_scan}} to produce a summary plot for.}

\item{...}{Additional parameters.}
}
\value{
A \code{\link[ggplot2]{ggplot2}} object for autoplot or nothing for plot, which prints the object directly.
}
\description{
Produce a summary plot for a \code{\link{deep_mutational_scan}} object, showing a heatmap with the ER score for
each substitution at each position.
}
\section{Functions}{
\itemize{
\item \code{plot.deep_mutational_scan}: S3 plot method
}}

