% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{dms_s3}
\alias{dms_s3}
\alias{format.deep_mutational_scan}
\alias{print.deep_mutational_scan}
\alias{str.deep_mutational_scan}
\alias{as_tibble.deep_mutational_scan}
\alias{as.data.frame.deep_mutational_scan}
\title{General S3 Methods for Deep Mutational Scans}
\usage{
\method{format}{deep_mutational_scan}(x, ...)

\method{print}{deep_mutational_scan}(x, ...)

\method{str}{deep_mutational_scan}(object, ...)

\method{as_tibble}{deep_mutational_scan}(x, ...)

\method{as.data.frame}{deep_mutational_scan}(x, ...)
}
\arguments{
\item{x, object}{\link{deep_mutational_scan} object}

\item{...}{Additional arguments}
}
\description{
General S3 Methods for Deep Mutational Scans
}
\section{Functions}{
\itemize{
\item \code{format.deep_mutational_scan}: S3 print method

\item \code{print.deep_mutational_scan}: S3 print method

\item \code{str.deep_mutational_scan}: S3 str method

\item \code{as_tibble.deep_mutational_scan}: S3 as_tibble method

\item \code{as.data.frame.deep_mutational_scan}: S3 as.data.frame method
}}

