% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise.R
\name{normalise_er}
\alias{normalise_er}
\title{Normalise deep mutational scanning fitness scores}
\usage{
normalise_er(x, q = 0.1)
}
\arguments{
\item{x}{Vector of fitness scores}

\item{q}{Quantile to normalise against. Values other than 0.1 differ from
those used in the combined dataset}
}
\description{
Normalise deep mutational scanning fitness scores
}
