% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_mavedb}
\alias{parse_mavedb}
\title{Parse data in the MaveDB format (called from parse_dms_data)}
\usage{
parse_mavedb(x, score_col = "score", ...)
}
\arguments{
\item{x}{A tibble.}

\item{score_col}{String. Column containing fitness scores.}

\item{...}{Ignored.}
}
\description{
Parse data in the MaveDB format (called from parse_dms_data)
}
