% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R
\name{parse_sequence}
\alias{parse_sequence}
\title{Parse data in the sequence format (called from parse_dms_data)}
\usage{
parse_sequence(x, ...)
}
\arguments{
\item{x}{A tibble.}

\item{...}{Ignored.}
}
\description{
Parse data in the sequence format (called from parse_dms_data)
}
