% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_cluster_frequencies}
\alias{plot_cluster_frequencies}
\title{Plot amino acid subtype frequencies}
\usage{
plot_cluster_frequencies(x)
}
\arguments{
\item{x}{\code{\link{deep_mutational_scan}} or \link[=rbind.deep_mutational_scan]{combined DMS data frame}.}
}
\description{
Plot amino acid subtype frequencies
}
