% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_landscape}
\alias{plot_landscape}
\title{Plot a new study on the deep mutational landscape}
\usage{
plot_landscape(x, feature = NULL)
}
\arguments{
\item{x}{\code{\link{deep_mutational_scan}}.}

\item{feature}{String name of a feature to project onto the background landscape. This can be a numeric column from
the \code{\link{deep_landscape}} dataset or the special values "count" or "subtype", which map the position density
and most common subtype category respectively.}
}
\value{
A \code{\link[ggplot2]{ggplot2}} plot.
}
\description{
Project new data onto the deep mutational landscape derived from the combined data, viewed in UMAP space. This view
allows you to determine if new data follows the expected distribution and compare it with the average properties for
positions in the same regions of the landscape. This can give insights into the likely properties of positions in the
new protein.
}
\details{
Points represent positions from the new protein(s), positioned based on predicted UMAP coordinates from the
combined landscape model. These are plotted on top of the whole combined landscape, which is represented as grey
points when no feature is used and coloured hexagonal bins when one is. These bins are coloured based on the mean
value of the chosen feature for all points that fall within the hexagon for numerical landscape features.
If the chosen feature is a FoldX Gibbs Free Energy prediction values are initially clamped to -5 <= x <= 5 because
larger values lead to difficult to read scales and are generally thought to represent program quirks rather than
real biology. There are also two special values for feature "count" and "cluster", which colour bins by the
number of landscape positions and the most common cluster type, respectively.

The input data is annotated using \code{\link{annotate}} if it is not already, based on the \code{annotated} flag
for deep_mutational_scan objects and the presence of the required \code{umap1/2} columns for data frames. It is
better to save the results of \code{annotate} if doing multiple downstream analyses because can be a relatively slow
operation.
}
\examples{
dms <- deepscanscape::deep_scans$p53

# Plot point positions
plot_landscape(dms)

# Plot against mean conservation
dms <- annotate(dms)
plot_landscape(dms, feature = "mean_sift")

# Plot against spceial features
plot_landscape(dms, feature = "count")
plot_landscape(dms, feature = "cluster")

# Plot multiple studies
comb_dms <- bind_scans(dms, annotate_missing = TRUE)
plot_landscape(comb_dms, feature = "total_energy")

}
