% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{rbind.deep_mutational_scan}
\alias{rbind.deep_mutational_scan}
\title{Combine deep mutational scan data}
\usage{
\method{rbind}{deep_mutational_scan}(...)
}
\arguments{
\item{...}{\code{\link{deep_mutational_scan}} objects to combine or lists of such objects.}
}
\value{
A \code{\link[tibble]{tibble}} whose rows contain positional data from all the provided scans.
}
\description{
Combine multiple \code{\link{deep_mutational_scan}} objects into a single \code{\link[tibble]{tibble}}, including
columns giving the study and gene of each position.
}
