% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recluster.R
\name{recluster_dms}
\alias{recluster_dms}
\title{Recluster DMS data with a combined dataset}
\usage{
recluster_dms(x)
}
\arguments{
\item{x}{\link{deep_mutational_scan} to recluster}
}
\description{
Recluster DMS data with a combined dataset
}
