% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{summary.deep_mutational_scan}
\alias{summary.deep_mutational_scan}
\title{Summarise Deep Mutational Scans}
\usage{
\method{summary}{deep_mutational_scan}(object, ...)
}
\arguments{
\item{object}{A \link{deep_mutational_scan} object.}

\item{...}{Additional arguments}
}
\description{
Summarise Deep Mutational Scans
}
