% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise.R
\name{transform_dms}
\alias{transform_dms}
\title{Again, could work on df or vector, probably df
Standardise DMS data}
\usage{
transform_dms(x, trans, base = 2)
}
\arguments{
\item{x}{Vector of fitness scores}

\item{trans}{Transform to apply (see description). Accepts either a
string or function}

\item{base}{Log base of current scores (used for log transform)}
}
\description{
Again, could work on df or vector, probably df
Standardise DMS data
}
