% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep_mutational_scan.R
\name{validate_deep_mutational_scan}
\alias{validate_deep_mutational_scan}
\title{Validate deep_mutational_scan objects}
\usage{
validate_deep_mutational_scan(x)
}
\arguments{
\item{x}{\code{\link{deep_mutational_scan}} object.}
}
\value{
The input is returned unaltered, since the main purpose of the function is to raise an error if the
  object is erroneous.
}
\description{
Check properties of deep_mutational_scan object adhere to various baseline expectations, which would lead to
downstream errors if not caught.
}
\details{
The following assumptions are currently checked:
\itemize{
  \item Correct class set
  \item Correct fields present
  \item \code{data} is a tibble
  \item \code{annotated} is logical
  \item \code{impute_matrix} is a tibble or NA
  \item \code{cluster} is a tibble or NA
  \item \code{data} contains the correct columns
  \item \code{data} contains no duplicate rows
  \item \code{impute_matrix} contains the correct type of values
  \item Annotated objects contain the correct \code{data} and \code{cluster} columns
}
}
